+++ lib/src/abstractions/service_locator_adapter.dart
@@ -16,7 +16,7 @@ abstract class ServiceLocatorAdapter {
   /// Asynchronously registers a transient factory for type [T].
   Future<Triplet<Success, Conflict, InternalError>>
       registerTransient<T extends Object>(
-    Future<T> Function(Map<String, dynamic> namedArgs) factory, {
+    Future<T> Function(Map<String, dynamic> args) factory, {
     required Set<Type>? interfaces,
     required String? name,
     required dynamic key,
@@ -40,7 +40,7 @@ abstract class ServiceLocatorAdapter {
     required String? name,
     required dynamic key,
     required String? environment,
-    required Map<String, dynamic> namedArgs,
+    required Map<String, dynamic> args,
     required bool resolveFirst,
   });
 
@@ -50,7 +50,7 @@ abstract class ServiceLocatorAdapter {
     required String? name,
     required dynamic key,
     required String? environment,
-    required Map<String, dynamic> namedArgs,
+    required Map<String, dynamic> args,
   });
 
   /// Asynchronously overrides an existing registration with a new `Singleton` of type [T].
@@ -66,7 +66,7 @@ abstract class ServiceLocatorAdapter {
   /// Asynchronously overrides an existing registration with a new `Transient` of type [T].
   Future<Doublet<Success, InternalError>>
       overrideWithTransient<T extends Object>(
-    Future<T> Function(Map<String, dynamic> namedArgs) factory, {
+    Future<T> Function(Map<String, dynamic> args) factory, {
     required Set<Type>? interfaces,
     required String? name,
     required dynamic key,