+++ test/named_test.dart
@@ -16,46 +16,46 @@ void main() {
   group('Concrete', () {
     test('Should Register', () async {
       // Arrange
-      final instance = NoDependencies();
+      final instance = SimpleService();
 
       // Act
       await ServiceLocator.I.registerSingleton(instance, name: 'one');
-      final resolvedInstance = await ServiceLocator.I.resolve<NoDependencies>();
+      final resolvedInstance = await ServiceLocator.I.resolve<SimpleService>();
 
       // Assert
-      expect(resolvedInstance, isA<NoDependencies>());
+      expect(resolvedInstance, isA<SimpleService>());
     });
 
     test('Different Names | Should Respect Names', () async {
       // Arrange
-      final instanceOne = NoDependencies();
-      final instanceTwo = NoDependencies();
+      final instanceOne = SimpleService();
+      final instanceTwo = SimpleService();
 
       // Act
       await ServiceLocator.I.registerSingleton(instanceOne, name: 'one');
       await ServiceLocator.I.registerSingleton(instanceTwo, name: 'two');
       final resolvedInstanceOne =
-          await ServiceLocator.I.resolve<NoDependencies>(name: 'one');
+          await ServiceLocator.I.resolve<SimpleService>(name: 'one');
       final resolvedInstanceTwo =
-          await ServiceLocator.I.resolve<NoDependencies>(name: 'two');
+          await ServiceLocator.I.resolve<SimpleService>(name: 'two');
 
       // Assert
-      expect(resolvedInstanceOne, isA<NoDependencies>());
-      expect(resolvedInstanceTwo, isA<NoDependencies>());
+      expect(resolvedInstanceOne, isA<SimpleService>());
+      expect(resolvedInstanceTwo, isA<SimpleService>());
     });
 
     test('One named Instance | Should Return One Instance', () async {
       // Arrange
-      final unnamedInstance = NoDependencies();
-      final namedInstance = NoDependencies();
+      final unnamedInstance = SimpleService();
+      final namedInstance = SimpleService();
 
       // Act
       await ServiceLocator.I.registerSingleton(unnamedInstance);
       await ServiceLocator.I.registerSingleton(namedInstance, name: 'named');
 
-      final allInstances = await ServiceLocator.I.resolveAll<NoDependencies>();
+      final allInstances = await ServiceLocator.I.resolveAll<SimpleService>();
       final namedInstances =
-          await ServiceLocator.I.resolveAll<NoDependencies>(name: 'named');
+          await ServiceLocator.I.resolveAll<SimpleService>(name: 'named');
 
       // Assert
       expect(allInstances, hasLength(2));
@@ -64,14 +64,14 @@ void main() {
 
     test('Multiple With Same Name | Should Resolve All', () async {
       // Arrange
-      final instanceOne = NoDependencies();
-      final instanceTwo = NoDependencies();
+      final instanceOne = SimpleService();
+      final instanceTwo = SimpleService();
 
       // Act
       await ServiceLocator.I.registerSingleton(instanceOne, name: 'instance');
       await ServiceLocator.I.registerSingleton(instanceTwo, name: 'instance');
       final instances =
-          await ServiceLocator.I.resolveAll<NoDependencies>(name: 'instance');
+          await ServiceLocator.I.resolveAll<SimpleService>(name: 'instance');
 
       // Assert
       expect(instances, hasLength(2));